[fix moderators related issues] update current mods
