+++ src/styles/Collapsible.css
@@ -5,6 +5,9 @@
     border-radius: 8px;
     border: 1px solid #b8b9dc;
     background: #f7f7fe;
+
+    overflow-y: overlay;
+    max-height: 250px;
   }
 
   .collapsibleHeaderContainer {
