+++ src/styles/ForumContent.css
@@ -142,6 +142,11 @@
     margin-bottom: 12px;
     width: 100%;
     font-family: Manrope, sans-serif;
+
+    ul {
+      max-height: 260px;
+      overflow-y: overlay;
+    }
   }
 
   .addModeratorsLabel {
